refactor: Move MySqlAdapter to External/Database folder
